Implemented Behavioural Design Pattern
Implemented Observer Pattern. This pattern observers key information in
the DoctoLib class such as patients, centers, and more. This information
is fetched from a JSON page and once it retreives it we notify the
listeners accordingly. There is a patient listener so whenever patients
are updated in DoctoLib the listeners are notified. We can then use this information in the Application class.